new logging UI
