+++ workoutApp_v1/workoutApp_v1/Views/Logging/PlayRoutine.swift
@@ -1,33 +1,37 @@
 //
-//  PlayRoutine.swift
+//  PageView.swift
 //  workoutApp_v1
 //
-//  Created by Kamiar Coffey on 10/30/19.
+//  Created by Kamiar Coffey on 11/9/19.
 //  Copyright © 2019 Kamiar Coffey. All rights reserved.
 //
-
 
 import Foundation
 import SwiftUI
 import Combine
 
-// TODO: Fix the alert button
-//        .navigationBarTitle(Text(routine.name), displayMode: .inline).onTapGesture {
-//            self.isShowingQuitAlert.toggle()
-//        }
-
-
-struct PlayRoutine: View {
+struct PlayRoutine : View {
+    let limit: Double = 15
+    let step: Double = 0.3
     
     var routine: PeleRoutine
-    @State private var isShowingQuitAlert = false
     @ObservedObject var playRoutineViewModel: PlayRoutineViewModel
+    @State private var isShowingQuitAlert = false
+    var pages: [LogExerciseViewModel]
     
     init(routine: PeleRoutine) {
         self.routine = routine
         self.playRoutineViewModel = PlayRoutineViewModel(with: routine)
+        self.pages = routine.exerciseList.map { LogExerciseViewModel(exercise: $0) }
     }
     
+    
+    @State var currentIndex = 0
+    @State var nextIndex = 1
+    
+    @State var progress: Double = 0
+    
+    @State private var shape = AnyView(Image(systemName: "play").foregroundColor(.blue).frame(width: 60.0, height: 60.0, alignment: .center))
     
     var body: some View {
         VStack {
@@ -39,18 +43,30 @@ struct PlayRoutine: View {
                     .padding(.all, 6)
                     .background(Color.orange)
             })
-            Divider()
-            ScrollView(.horizontal, content: {
-                HStack{
-                    ForEach(playRoutineViewModel.exercisesInProgress, id: \.id) { exerciseViewModel in
-                        LogExercise(logExerciseViewModel: exerciseViewModel, completedExercise: { newExercise in
-                            self.playRoutineViewModel.addExercise(new: newExercise)
-                        })
-                    }
-                    .padding(.leading, 10)
-                }
-            })
+            ZStack {
+                LogExercise(logExerciseViewModel: pages[currentIndex], completedExercise: { newExercise in
+                    self.playRoutineViewModel.addExercise(new: newExercise)
+                    self.refreshAnimatingViews()
+                })
+                    .offset(x: -CGFloat(pow(2, self.progress)))
+                LogExercise(logExerciseViewModel: pages[nextIndex], completedExercise: { newExercise in
+                    self.playRoutineViewModel.addExercise(new: newExercise)
+                    self.refreshAnimatingViews()
+                })
+                    .offset(x: CGFloat(pow(2, (self.limit - self.progress))))
+            }.edgesIgnoringSafeArea(.vertical)
+        }
+    }
+    
+    func refreshAnimatingViews() {
+        currentIndex = nextIndex
+        if nextIndex + 1 < pages.count {
+            nextIndex += 1
+        } else {
+            nextIndex = 0
         }
     }
 }
+
+
